+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import ChatApp from "./components/ChatApp/index.js";
-import SignIn from "./components/SignIn/index.js";
+import Login from "./components/Login/index.js";
 import SignUp from "./pages/SignUp/index.js";
 import "bootstrap/dist/css/bootstrap.min.css";
 // import axios from 'axios'
@@ -9,7 +9,6 @@ import "bootstrap/dist/css/bootstrap.min.css";
 // import "./App.css";
 
 function App() {
-
   // const [state, dispatch] = useGlobalContext()
 
   // useEffect(()=>{
@@ -23,9 +22,9 @@ function App() {
   //   // is there a user?
   //   if (user) {
   //     // put that user in the state
-  //     dispatch({ 
-  //       type: "LOGIN", 
-  //       email: user.email, 
+  //     dispatch({
+  //       type: "LOGIN",
+  //       email: user.email,
   //       apiToken: user.token
   //     });
   //   }
@@ -51,10 +50,10 @@ function App() {
 
   return (
     <Router>
-    <div className="App">
-      {/* <h1>Let's build AllSpeak</h1> */}
-      <Route exact path="/" component={SignIn} />
-        <Route exact path="/signin" component={SignIn} />
+      <div className="App">
+        {/* <h1>Let's build AllSpeak</h1> */}
+        <Route exact path="/" component={Login} />
+        <Route exact path="/login" component={Login} />
         <Route exact path="/signup" component={SignUp} />
 
         {/* full chat window with sidebar component */}
